ある程度の警告を消した

+++ NewEngine/Render/ScissorRectangle.cpp
@@ -11,9 +11,9 @@ void ScissorRectangle::Update()
 	// シザー矩形
 	D3D12_RECT scissorRect{};
 	scissorRect.left = 0; // 切り抜き座標左
-	scissorRect.right = scissorRect.left + renderWindow->GetWindowSize().x; // 切り抜き座標右
+	scissorRect.right = scissorRect.left + (LONG)renderWindow->GetWindowSize().x; // 切り抜き座標右
 	scissorRect.top = 0; // 切り抜き座標上
-	scissorRect.bottom = scissorRect.top + renderWindow->GetWindowSize().y; // 切り抜き座標下
+	scissorRect.bottom = scissorRect.top + (LONG)renderWindow->GetWindowSize().y; // 切り抜き座標下
 	// シザー矩形設定コマンドを、コマンドリストに積む
 	renderBase->GetCommandList()->RSSetScissorRects(1, &scissorRect);
 }
